# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AcronIQ Forge API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Add a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Remove a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /projects/{id}/scan:
    post:
      operationId: triggerScan
      tags: [projects]
      summary: Trigger a project scan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "202":
          description: Scan queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /projects/{id}/summary:
    get:
      operationId: getProjectSummary
      tags: [projects]
      summary: Get health summary for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSummary"
...
  /devices:
    get:
      operationId: listDevices
      tags: [devices]
      summary: List connected Forge Seed devices
      responses:
        "200":
          description: Devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
    post:
      operationId: registerDevice
      tags: [devices]
      summary: Register a new device (generates pairing token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceInput"
      responses:
        "201":
          description: Device registered with pairing token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /devices/{id}:
    get:
      operationId: getDevice
      tags: [devices]
      summary: Get a device
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
    delete:
      operationId: deleteDevice
      tags: [devices]
      summary: Unlink a device
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /devices/{id}/heartbeat:
    patch:
      operationId: deviceHeartbeat
      tags: [devices]
      summary: Device sends heartbeat with status update
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceHeartbeat"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
...
  /memory:
    get:
      operationId: listMemoryEntries
      tags: [memory]
      summary: List engineering memory entries
      parameters:
        - name: projectId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: category
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Memory entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryEntry"
    post:
      operationId: createMemoryEntry
      tags: [memory]
      summary: Add a memory entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryEntry"
  /memory/{id}:
    patch:
      operationId: updateMemoryEntry
      tags: [memory]
      summary: Update a memory entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryEntryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryEntry"
    delete:
      operationId: deleteMemoryEntry
      tags: [memory]
      summary: Delete a memory entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
...
  schemas:
...
    Project:
      type: object
      required: [id, name, repoUrl, status, healthScore, language, branch, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        repoUrl:
          type: string
        status:
          type: string
        healthScore:
          type: integer
        language:
          type: string
        branch:
          type: string
        lastScanAt:
          type: ["string", "null"]
        createdAt:
          type: string
    ProjectInput:
      type: object
      required: [name, repoUrl]
      properties:
        name:
          type: string
          minLength: 1
        repoUrl:
          type: string
          minLength: 1
        language:
          type: string
        branch:
          type: string
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        repoUrl:
          type: string
        language:
          type: string
        branch:
          type: string
        status:
          type: string
    ProjectSummary:
      type: object
      required: [projectId, healthScore, openIssues, criticalIssues, lastScanAt, filesIndexed, openTasks, openPullRequests]
      properties:
        projectId:
          type: integer
        healthScore:
          type: integer
        openIssues:
          type: integer
        criticalIssues:
          type: integer
        lastScanAt:
          type: ["string", "null"]
        filesIndexed:
          type: integer
        openTasks:
          type: integer
        openPullRequests:
          type: integer
...
    Task:
      type: object
      required: [id, projectId, type, status, prompt, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        projectName:
          type: string
        type:
          type: string
        status:
          type: string
        prompt:
          type: string
        plan:
          type: ["string", "null"]
        filesModified:
          type: array
          items:
            type: string
        buildStatus:
          type: ["string", "null"]
        confidenceScore:
          type: ["integer", "null"]
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
...
    Device:
      type: object
      required: [id, userId, name, platform, status, pairingToken, createdAt, ollamaAvailable]
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        platform:
          type: string
          description: "windows | macos | linux"
        status:
          type: string
          description: "online | offline | idle"
        pairingToken:
          type: string
        agentVersion:
          type: ["string", "null"]
        ollamaAvailable:
          type: boolean
        ollamaVersion:
          type: ["string", "null"]
        activeModel:
          type: ["string", "null"]
        lastHeartbeatAt:
          type: ["string", "null"]
        createdAt:
          type: string
    DeviceInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        platform:
          type: string
    DeviceHeartbeat:
      type: object
      properties:
        status:
          type: string
        ollamaAvailable:
          type: boolean
        ollamaVersion:
          type: ["string", "null"]
        activeModel:
          type: ["string", "null"]
        agentVersion:
          type: ["string", "null"]
...
    MemoryEntry:
      type: object
      required: [id, userId, category, key, value, source, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        projectId:
          type: ["integer", "null"]
        category:
          type: string
          description: "architecture | fix | convention | preference | workflow"
        key:
          type: string
        value:
          type: string
        source:
          type: string
          description: "manual | agent | scan"
        createdAt:
          type: string
        updatedAt:
          type: string
    MemoryEntryInput:
      type: object
      required: [category, key, value]
      properties:
        category:
          type: string
          minLength: 1
        key:
          type: string
          minLength: 1
        value:
          type: string
          minLength: 1
        projectId:
          type: ["integer", "null"]
        source:
          type: string
    MemoryEntryUpdate:
      type: object
      properties:
        value:
          type: string
        category:
          type: string
...