# packages/db/src/schema/tasks.ts
import { pgTable, serial, text, integer, timestamp, json } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const tasksTable = pgTable("tasks", {
  id: serial("id").primaryKey(),
  projectId: integer("project_id").notNull(),
  type: text("type").notNull().default("fix"),
  status: text("status").notNull().default("pending"),
  prompt: text("prompt").notNull(),
  plan: text("plan"),
  filesModified: json("files_modified").$type<string[]>().notNull().default([]),
  buildStatus: text("build_status"),
  confidenceScore: integer("confidence_score"),
  createdAt: timestamp("created_at").notNull().defaultNow(),
  completedAt: timestamp("completed_at"),
});

export const insertTaskSchema = createInsertSchema(tasksTable).omit({ id: true, createdAt: true });
export type InsertTask = z.infer<typeof insertTaskSchema>;
export type Task = typeof tasksTable.$inferSelect;